Create highly_class-ical_module.py
its highly class-ical, get it? Because there are so many classes? That was my best try at a pun. Also, take a close look at the code. Now, if you are a python programmer; PLEASE FORGIVE ME I had to do this. Its a new... how do I describe this... Its the amber heard of python programs. No, it's not April 1st.